BDS-118,119: Medical service delete.

+++ bsdclinic-web/src/main/resources/static/admin/custom/js/medicine/script.js
@@ -207,8 +207,10 @@ export const MedicineDeletion = (function () {
     const handleShowMedicineDeletionConfirmation = () => {
         MedicineList.medicineListTableSelector.on('click', '.show-deletion-confirmation-btn', function () {
             App.showSweetAlertConfirmation('error', confirmApplyTitle, cannotRedoAfterDeleting).then(() => {
-                const rowData = MedicineList.medicineListTableSelector.DataTable().row($(this).closest('tr')).data();
-                deleteMedicine(rowData.medicineId);
+                if(result.isConfirmed) {
+                    const rowData = MedicineList.medicineListTableSelector.DataTable().row($(this).closest('tr')).data();
+                    deleteMedicine(rowData.medicineId);
+                }
             });
         });
     }
